Moved the method to tools
